1) Ok, как раньше только все сделал как просил - структура та же, что и в твоей ветке plot2movie.
