+++ models/ner/char.dict
@@ -1,0 +1,47 @@
+e	124062
+a	110230
+t	96490
+i	91878
+o	76612
+r	73378
+s	72518
+n	68969
+h	60847
+l	43284
+d	42412
+m	41386
+c	33422
+w	28267
+f	24774
+g	21944
+u	21733
+y	19308
+v	18438
+b	17316
+p	15896
+k	9448
+1	5415
+0	4920
+9	3798
+j	3297
+2	2387
+x	1244
+z	1212
+3	923
+8	869
+7	815
+q	638
+6	599
+5	566
+4	518
+T	2
+<	1
+D	1
+O	1
+C	1
+S	1
+A	1
+R	1
+>	1
+,	1
+$	1
