+++ models/ner/tag.dict
@@ -1,0 +1,49 @@
+O	116900
+I-Plot	62095
+B-Plot	6467
+I-Actor	6120
+B-Actor	5009
+B-GENRE	4354
+I-TITLE	3495
+I-ACTOR	3474
+B-Genre	3384
+I-Origin	3335
+B-ACTOR	3220
+B-YEAR	2858
+B-Year	2702
+I-YEAR	2456
+B-TITLE	2376
+I-Genre	2283
+B-RATING	2007
+B-PLOT	1927
+B-RATINGS_AVERAGE	1869
+I-DIRECTOR	1850
+B-Director	1787
+B-DIRECTOR	1720
+I-PLOT	1687
+I-RATINGS_AVERAGE	1673
+I-Director	1653
+I-Relationship	1206
+B-Character_Name	1025
+I-RATING	840
+I-Quote	817
+B-Opinion	810
+I-GENRE	786
+B-Origin	778
+I-Character_Name	760
+I-Award	719
+B-Relationship	580
+I-Opinion	539
+I-SONG	446
+B-CHARACTER	385
+I-CHARACTER	342
+B-Award	309
+B-SONG	245
+B-REVIEW	221
+I-Year	195
+I-Soundtrack	158
+I-REVIEW	132
+B-Quote	126
+B-TRAILER	113
+B-Soundtrack	50
+I-TRAILER	7
