+++ models/ner_config.json
@@ -1,0 +1,161 @@
+{
+    "dataset_reader": {
+      "class_name": "conll2003_reader",
+      "data_path": "./data/ner/"
+    },
+    "dataset_iterator": {
+      "class_name": "data_learning_iterator",
+      "seed": 42
+    },
+    "chainer": {
+      "in": ["x"],
+      "in_y": ["y"],
+      "pipe": [
+        {
+          "in": ["x"],
+          "class_name": "lazy_tokenizer",
+          "out": ["x_tokens"]
+        },
+        {
+          "in": ["x_tokens"],
+          "class_name": "str_lower",
+          "out": ["x_lower"]
+        },
+        {
+          "in": ["x_lower"],
+          "class_name": "sanitizer",
+          "nums": true,
+          "out": ["x_san"]
+        },
+        {
+          "in": ["y"],
+          "id": "tag_vocab",
+          "class_name": "simple_vocab",
+          "pad_with_zeros": true,
+          "fit_on": ["y"],
+          "save_path": "{MODELS_PATH}/ner/tag.dict",
+          "load_path": "{MODELS_PATH}/ner/tag.dict",
+          "out": ["y_ind"]
+        },
+        {
+          "in": ["x_tokens"],
+          "class_name": "char_splitter",
+          "out": ["x_char"]
+        },
+        {
+          "in": ["x_char"],
+          "id": "char_vocab",
+          "class_name": "simple_vocab",
+          "pad_with_zeros": true,
+          "fit_on": ["x_char"],
+          "save_path": "{MODELS_PATH}/ner/char.dict",
+          "load_path": "{MODELS_PATH}/ner/char.dict",
+          "out": ["x_char_ind"]
+        },
+        {
+          "in": ["x_tokens"],
+          "class_name": "mask",
+          "out": ["mask"]
+        },
+        {
+          "in": ["x_san"],
+          "id": "glove_emb",
+          "class_name": "glove",
+          "pad_zero": true,
+          "load_path": "{DOWNLOADS_PATH}/embeddings/glove.6B.100d.txt",
+  
+          "out": ["x_emb"]
+        },
+        {
+          "id": "embeddings_char",
+          "class_name": "emb_mat_assembler",
+          "character_level": true,
+          "emb_dim": 32,
+          "embedder": "#glove_emb",
+          "vocab": "#char_vocab"
+        },
+        {
+          "id": "capitalization",
+          "class_name": "capitalization_featurizer",
+          "in": ["x_tokens"],
+          "out": ["cap"]
+        },
+        {
+          "in": ["x_emb", "mask", "x_char_ind", "cap"],
+          "in_y": ["y_ind"],
+          "out": ["y_predicted"],
+          "class_name": "ner",
+          "main": true,
+          "token_emb_dim": "#glove_emb.dim",
+          "n_hidden_list": [256, 256, 256],
+          "net_type": "rnn",
+          "cell_type": "lstm",
+          "use_cudnn_rnn": true,
+          "n_tags": "#tag_vocab.len",
+          "capitalization_dim": "#capitalization.dim",
+          "char_emb_dim": "#embeddings_char.dim",
+          "save_path": "{MODELS_PATH}/ner/model",
+          "load_path": "{MODELS_PATH}/ner/model",
+          "char_emb_mat": "#embeddings_char.emb_mat",
+          "two_dense_on_top": true,
+          "use_crf": true,
+          "use_batch_norm": true,
+          "embeddings_dropout": true,
+          "top_dropout": true,
+          "intra_layer_dropout": false,
+          "l2_reg": 0,
+          "learning_rate": 3e-3,
+          "dropout_keep_prob": 0.7
+        },
+        {
+          "ref": "tag_vocab",
+          "in": ["y_predicted"],
+          "out": ["tags"]
+        }
+      ],
+  
+      "out": ["x_tokens", "tags"]
+    },
+    "train": {
+      "epochs": 100,
+      "batch_size": 64,
+  
+      "metrics": [
+        {
+          "name": "ner_f1",
+          "inputs": ["y", "tags"]
+        }
+      ],
+      "validation_patience": 7,
+      "val_every_n_epochs": 1,
+  
+      "log_every_n_epochs": -1,
+      "show_examples": false
+    },
+    "metadata": {
+      "variables": {
+        "ROOT_PATH": "./models/",
+        "DOWNLOADS_PATH": "{ROOT_PATH}/downloads",
+        "MODELS_PATH": "{ROOT_PATH}/"
+      },
+      "requirements": [
+        "{DEEPPAVLOV_PATH}/requirements/gensim.txt",
+        "{DEEPPAVLOV_PATH}/requirements/tf.txt"
+      ],
+      "labels": {
+        "telegram_utils": "NERCoNLL2003Model",
+        "server_utils": "NER"
+      },
+      "download": [
+        {
+          "url": "http://files.deeppavlov.ai/deeppavlov_data/ner_ontonotes_v3_cpu_compatible.tar.gz",
+          "subdir": "{MODELS_PATH}"
+        },
+        {
+          "url": "http://files.deeppavlov.ai/embeddings/glove.6B.100d.txt",
+          "subdir": "{DOWNLOADS_PATH}/embeddings"
+        }
+      ]
+    }
+  
+  }
